+++ examples/MicroserviceExample/WebApi/Startup.cs
@@ -50,8 +50,10 @@ namespace WebApi
                 }));
         }
 
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+        public void Configure(IApplicationBuilder app, IHost host, IWebHostEnvironment env)
         {
+            host.UseOpenTelemetry();
+
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();